pyproject.toml
[project]
name = "video-upscale"
version = "1.0.0"
description = "2x/4x Video Super-Resolution with PureVision/PureScale ONNX + TensorRT"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "brucx" }
]
keywords = ["video", "upscale", "super-resolution", "onnx", "tensorrt", "gpu"]

dependencies = [
    "numpy>=1.24.0",
    "onnxruntime-gpu>=1.16.0",
    "pycuda>=2024.1",
    "PyNvVideoCodec>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
]

[project.scripts]
video-upscale = "upscale:main"

[project.urls]
Homepage = "https://github.com/brucx/video-upscale"
Repository = "https://github.com/brucx/video-upscale"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
